+++ pac4j-javaee/src/main/java/org/pac4j/jee/context/JEEContext.java
@@ -251,4 +251,10 @@ public class JEEContext implements WebContext {
     public String getProtocol() {
         return request.getProtocol();
     }
+
+    /** {@inheritDoc} */
+    @Override
+    public Optional<String> getQueryString() {
+        return Optional.ofNullable(request.getQueryString());
+    }
 }