+++ documentation/docs/release-notes.md
@@ -28,6 +28,8 @@ title: Release notes&#58;
 - Created the `SpringResourceLoader` for OIDC/SAML metadata loading: for the OIDC support, the `discoveryURI` can use the "file:", "classpath:" or "resource:" prefix in addition to HTTP/HTTPS URLs
 - The `DefaultSessionLogoutHandler` smartly tries a front channel logout and then a back channel logout
 - The `OidcProfile` will internally encode/decode codes, access and refresh tokens. Asking the profile to return back the actual object will effectively reconstruct it, to avoid  issues with JSON serialization.
+- Added `getQueryString` on the `WebContext`
+
 ---
 
 ### JDK11: